requirements.txt
fastapi
jinja2
python-multipart
email-validator
python-dotenv
uvicorn
pydantic[email]
